refactor(dto): move LottoCountsDto from domain to dto package
