+++ src/main/java/dto/LottoCountsDto.java
@@ -1,4 +1,4 @@
-package domain;
+package dto;
 
 public class LottoCountsDto {
